chore: move useDownloadClick into hooks folder
